requirements.txt
fastapi
uvicorn
requests
Pillow
numpy
shapely
torch
diffusers[torch]
transformers
accelerate
opencv-python-headless